requirements.txt
opencv-python
pyyaml
typeguard
Flask
google-genai
ytmusicapi